requirements.txt
paramiko==2.10.3
python-nmap==0.7.1
requests==2.28.1
tqdm==4.64.0
gnupg==2.3.1
beautifulsoup4==4.11.1
lxml==4.9.1
openvas_lib==1.1.4
